Refactor(Domain, Test): 닉네임 validation 코드 리팩토링
- 더 자세한 예외 발생
- 예외에서 상황에 맞는 커스텀 ErrorCode 세팅
