+++ src/main/java/podo/odeego/domain/member/exception/MemberNicknameOutOfBoundsException.java
@@ -1,0 +1,11 @@
+package podo.odeego.domain.member.exception;
+
+import podo.odeego.global.error.ErrorCode;
+import podo.odeego.global.error.exception.InvalidValueException;
+
+public class MemberNicknameOutOfBoundsException extends InvalidValueException {
+
+	public MemberNicknameOutOfBoundsException(String message) {
+		super(message, ErrorCode.MEMBER_NICKNAME_OUT_OF_BOUNDS);
+	}
+}
